pyproject.toml
[tool.poetry]
name = "Kaguya"
version = "0.1.0"
description = "Kaguya"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.3.2"
nonebot-plugin-apscheduler = "^0.1.2"
pillow = "^8.1.0"
ujson = "^4.0.2"
nonebot-plugin-test = "^0.2.0"
zhconv = "^1.4.1"
bs4 = "^0.0.1"
aiocqhttp = "^1.3.0"
matplotlib = "^3.3.4"
pygtrie = "^2.4.2"
requests = "^2.25.1"
fuzzywuzzy = "^0.18.0"
lxml = "^4.6.2"
sogou-tr-free = "^0.0.11"
tinydb = "^4.3.0"
quart-cors = "^0.3.0"
flask = "^1.1.2"
flask-cors = "^3.0.10"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true